feat: 승패계산에서 Running / Finished abstract 추가 및 패키지 분리
