+++ src/main/kotlin/blackjack/domain/gamestate/Competition.kt
@@ -1,0 +1,7 @@
+package blackjack.domain.gamestate
+
+enum class Competition {
+    WIN,
+    LOSE,
+    DRAW,
+}
