+++ src/main/kotlin/blackjack/domain/gamestate/GameState.kt
@@ -1,6 +1,7 @@
 package blackjack.domain.gamestate
 
 import blackjack.domain.card.Card
+import blackjack.domain.gamestate.finished.Stay
 
 interface GameState {
 
@@ -13,4 +14,6 @@ interface GameState {
     fun isBust(): Boolean
 
     fun score(): Int
+
+    fun compete(gameState: GameState): Competition
 }
