+++ src/main/kotlin/blackjack/domain/gamestate/finished/Stay.kt
@@ -1,11 +1,12 @@
-package blackjack.domain.gamestate
+package blackjack.domain.gamestate.finished
 
-import blackjack.domain.card.Card
 import blackjack.domain.card.Cards
+import blackjack.domain.gamestate.Competition
+import blackjack.domain.gamestate.GameState
 
 class Stay(
     val cards: Cards,
-) : GameState {
+) : Finished() {
     init {
         require(cards.isInitialHand().not()) { "2장 미만의 카드로 생성될 수 없다." }
         require(cards.isBust().not()) { "버스트 카드로 생성될 수 없다." }
@@ -13,11 +14,11 @@ class Stay(
 
     override fun cards() = cards.values
 
-    override fun draw(card: Card) = throw IllegalStateException("종료된 게임은 draw할 수 없다.")
-
-    override fun stay() = throw IllegalStateException("종료된 게임은 stay할 수 없다.")
-
     override fun isBust() = false
 
     override fun score() = cards.score()
+
+    override fun compete(gameState: GameState): Competition {
+        TODO("Not yet implemented")
+    }
 }
