+++ src/main/kotlin/blackjack/domain/gamestate/running/InitialHand.kt
@@ -1,11 +1,12 @@
-package blackjack.domain.gamestate
+package blackjack.domain.gamestate.running
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.Cards
+import blackjack.domain.gamestate.GameState
 
 class InitialHand(
     val cards: Cards = Cards(),
-) : GameState {
+) : Running() {
     init {
         require(cards.isInitialHand()) { "초기 핸드는 2장 이상 가질 수 없다." }
     }
@@ -23,6 +24,4 @@ class InitialHand(
     override fun stay() = throw IllegalStateException("2장을 받기전에는 카드를 그만받을 수 없다.")
 
     override fun isBust() = false
-
-    override fun score() = throw IllegalStateException("턴이 종료되지 않아 점수를 반환할 수 없다.")
 }
